genes/human/RBFOX3/RBFOX3-bioinformatics/rbfox3_analysis/pyproject.toml
[project]
name = "rbfox3-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "biopython>=1.85",
    "matplotlib>=3.10.5",
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
]
